Permalinks trailing slash fix
Jekyll’s page.url always starts with a leading slash, creating issues
when combining it with baseurl. Best way to fix I think is to change
_config to remove trailing slash from baseurl, then adding it only as
it is needed in templates. Another option would be to filter the
baseurl in times when it would create double-slash. Went with the
former.